+++ lib/gs1.rb
@@ -1,21 +1,3 @@
-require 'gs1/version'
-
-require 'gs1/ai'
-require 'gs1/check_digit_calculator'
-require 'gs1/extensions'
-require 'gs1/definitions'
-require 'gs1/validations'
-
-require 'gs1/record'
-require 'gs1/batch'
-require 'gs1/content'
-require 'gs1/expiration_date'
-require 'gs1/gtin'
-require 'gs1/serial_number'
-require 'gs1/sscc'
-
-require 'gs1/barcode'
-
 # GS1 module.
 #
 module GS1
@@ -36,16 +18,27 @@ module GS1
   # Configuration holds custom configuration parameters.
   #
   class Configuration
-    attr_accessor :company_prefix
+    attr_accessor :company_prefix, :ignore_extra_barcode_elements
     attr_writer :barcode_separator
+
+    def initialize
+      @ignore_extra_barcode_elements = true
+    end
 
     def barcode_separator
       @barcode_separator || GS1::Barcode::DEFAULT_SEPARATOR
     end
   end
 
-  AI_CLASSES = GS1::Record.descendants.each_with_object({}) do |klass, hash|
-    hash[klass.ai] = klass
+  def self.ai_classes
+    @ai_classes ||= begin
+      GeneratedAIClasses.ai_classes
+      # sort to get non-generated classes first
+      ai_classes = GS1::Record.descendants.sort_by { _1.generated ? 1 : 0 }
+      ai_classes.each_with_object({}) do |klass, hash|
+        hash[klass.ai] ||= klass
+      end
+    end
   end
 
   module AIDCMarketingLevels
@@ -54,3 +47,22 @@ module GS1
            HIGHEST = 3].freeze
   end
 end
+
+require 'gs1/version'
+
+require 'gs1/ai'
+require 'gs1/check_digit_calculator'
+require 'gs1/extensions'
+require 'gs1/definitions'
+require 'gs1/validations'
+
+require 'gs1/record'
+require 'gs1/generated_ai_classes'
+require 'gs1/batch'
+require 'gs1/content'
+require 'gs1/expiration_date'
+require 'gs1/gtin'
+require 'gs1/serial_number'
+require 'gs1/sscc'
+
+require 'gs1/barcode'