+++ lib/gs1/barcode/attribute_validators/ignoring_record_validator.rb
@@ -1,0 +1,11 @@
+module GS1
+  module Barcode
+    module AttributeValidators
+      class IgnoringRecordValidator < RecordValidator
+        private
+
+        def on_error(_barcode, _attribute_name); end
+      end
+    end
+  end
+end