+++ spec/gs1/barcode/base_spec.rb
@@ -46,11 +46,13 @@ RSpec.describe GS1::Barcode::Base do
     end
 
     context 'with barcode containing unknown application identifier' do
+      subject { dummy.from_scan(data, ai_classes: {}) }
+
       let(:data) do
         "#{GS1::AI::FISHING_GEAR_TYPE}00000000000000"
       end
 
-      it 'does not raise anyting' do
+      it 'does not raise anything' do
         expect { subject }.not_to raise_error
       end
 
@@ -62,19 +64,52 @@ RSpec.describe GS1::Barcode::Base do
         expect(subject.sscc).to be_nil
       end
     end
+
+    context 'when the barcode contains extra elements' do
+      subject { dummy.from_scan(data, separator: separator) }
+      let(:separator) { "\u001E" }
+
+      let(:data) do
+        '70040000' \
+          "#{GS1::ExpirationDate::AI}181016" \
+          "#{GS1::SerialNumber::AI}zzzzzzz#{separator}" \
+          "#{GS1::Batch::AI}THISISABATCH#{separator}" \
+          "#{GS1::SSCC::AI}123123123123123123"
+      end
+
+      it 'sets the known attributes' do
+        expect(subject).to have_attributes(
+          batch: GS1::Batch.new('THISISABATCH'),
+          expiration_date: GS1::ExpirationDate.new('181016'),
+          serial_number: GS1::SerialNumber.new('zzzzzzz'),
+          sscc: GS1::SSCC.new('123123123123123123'),
+          gtin: be_nil
+        )
+      end
+    end
   end
 
   describe '.from_scan!' do
-    subject { dummy.from_scan!(data) }
-
-    context 'with barcode containing unknown application identifier' do
-      let(:data) do
-        "#{GS1::AI::FISHING_GEAR_TYPE}00000000000000"
+    context 'when configured to not ignore extra barcode elements' do
+      around do |spec|
+        old_value = GS1.configuration.ignore_extra_barcode_elements
+        GS1.configuration.ignore_extra_barcode_elements = false
+        spec.run
+      ensure
+        GS1.configuration.ignore_extra_barcode_elements = old_value
       end
 
-      it 'raises invalid token error' do
-        expect { subject }.to raise_error(GS1::Barcode::InvalidTokenError,
-                                          'Unable to retrieve record from application identifier(s) 70, 700, 7009')
+      context 'with barcode containing unknown application identifier' do
+        subject { dummy.from_scan!(data, ai_classes: {}) }
+
+        let(:data) do
+          "#{GS1::AI::FISHING_GEAR_TYPE}00000000000000"
+        end
+
+        it 'raises invalid token error' do
+          expect { subject }.to raise_error(GS1::Barcode::InvalidTokenError,
+                                            'Unable to retrieve record from application identifier(s) 70, 700, 7009')
+        end
       end
     end
   end
@@ -117,7 +152,7 @@ RSpec.describe GS1::Barcode::Base do
     end
 
     context 'with barcode containing unknown application identifier' do
-      subject { dummy.scan_to_params(data) }
+      subject { dummy.scan_to_params(data, ai_classes: {}) }
       let(:data) { '123456' }
 
       it 'returns an empty array' do
@@ -128,7 +163,7 @@ RSpec.describe GS1::Barcode::Base do
 
   describe '.scan_to_params!' do
     context 'with barcode containing unknown application identifier' do
-      subject { dummy.scan_to_params!(data) }
+      subject { dummy.scan_to_params!(data, ai_classes: {}) }
       let(:data) { '123456' }
 
       it 'raises invalid token error' do
@@ -144,6 +179,49 @@ RSpec.describe GS1::Barcode::Base do
       it 'raises invalid token error' do
         expect { subject }.to raise_error(GS1::Barcode::InvalidTokenError,
                                           'Unable to retrieve data to GS1::Batch')
+      end
+    end
+  end
+
+  describe 'errors' do
+    subject(:base) { dummy.from_scan(data) }
+
+    context 'when the data is valid' do
+      let(:data) { "#{GS1::GTIN::AI}00000000000000" }
+      let(:dummy) { Class.new(GS1::Barcode::Base) { define_records GS1::GTIN } }
+
+      it 'returns no errors' do
+        expect(base.errors).to be_empty
+      end
+    end
+
+    context 'when the data is not valid' do
+      let(:data) { "#{GS1::GTIN::AI}00000000000000" }
+      let(:dummy) { Class.new(GS1::Barcode::Base) { define_records GS1::SSCC } }
+      let(:ignore_extra_barcode_elements) { nil }
+
+      around do |spec|
+        old_value = GS1.configuration.ignore_extra_barcode_elements
+        GS1.configuration.ignore_extra_barcode_elements = ignore_extra_barcode_elements
+        spec.run
+      ensure
+        GS1.configuration.ignore_extra_barcode_elements = old_value
+      end
+
+      context 'when configured to not ignore extra barcode elements' do
+        let(:ignore_extra_barcode_elements) { false }
+
+        it 'returns a list of errors' do
+          expect(base.errors).not_to be_empty
+        end
+      end
+
+      context 'when configured to not ignore extra barcode elements' do
+        let(:ignore_extra_barcode_elements) { true }
+
+        it 'returns a list of errors' do
+          expect(base.errors).to be_empty
+        end
       end
     end
   end